+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## Added
 - `Trial` a wrapper class for high-level usage, which runs the optimiser, evaluates the objective
  by scheduling jobs, updates the optimiser and summarises the results.
+- a `Job` from a script with command line arguments can now be run with 
+ named arguments passed as a dictionary instead of a tuple.  
 
 ## [0.3.1] - 2019-09-10
 ## Fixed